requirements.txt
requests
pandas
langchain
langchain-community
langchain-classic
langchain-google-genai
python-dotenv
